seclude key input
